Use Sort and httpPatternTrie
